requirements.txt
flask
azure-cognitiveservices-language-textanalytics
